Fixes to give the same time from the `/status` and the websocket
previously one was returning local and the other GMT. Now they both return GMT

Fixes #467
